Rename module to atomicswaptypes
